refactor(network): extract validateFilterOption helper
network/list.ts and network/har.ts both inlined the same validate-or-
throw block for the filter DSL. Pull it into network/shared.ts so each
caller becomes a single line. Refs #223.
